requirements.txt
# A 股量化金融平台核心依赖

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 数据获取
akshare>=1.12.0
tushare>=1.2.0

# 可视化
matplotlib>=3.7.0
plotly>=5.18.0

# 统计分析
scipy>=1.11.0
scikit-learn>=1.3.0

# Jupyter
jupyter>=1.0.0
notebook>=7.0.0

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 日志
colorlog>=6.8.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0

# 代码质量
black>=23.0.0
flake8>=7.0.0
mypy>=1.8.0
pre-commit>=3.6.0

# 性能优化
numba>=0.58.0

# 日期处理
python-dateutil>=2.8.0

# 请求库
requests>=2.31.0

# 其他工具
tqdm>=4.66.0  # 进度条
